Testing ds-guard

+++ src/token.sol
@@ -58,7 +58,7 @@ contract DSToken is DSTokenBase(0), DSAuth {
         _supply -= x;
     }
     
-    function mint(uint x) auth stoppable {
+    function mint(uint256 x) auth stoppable {
         assert(_balances[msg.sender] + x >= _balances[msg.sender]);
         _balances[msg.sender] += x;
         _supply += x;
